split RotationAccumulator into its own files
